Adding Vibe.d-based website.
Most of the work is not apparent here, and was done in Brian Schott's libddoc library.
The goal is to be able to keep the content purely in DDOC, thus allowing us to generate the
other targets (pdf, mobi) the same way we did (but with a D-backed solution), be at the same
time we'd be able to write richer web page thanks to Vibe.d.
As a side effect, it has the potential to kill the Makefiles on the long run, and resort solely on dub for building the site.